test: 청원 Service 단위/통합 테스트로 구조 변경 #27
