+++ src/test/kotlin/com/example/echo/domain/petition/service/PetitionServiceUnitTests.kt
@@ -3,7 +3,6 @@ package com.example.echo.domain.petition.service
 import com.example.echo.domain.member.entity.Member
 import com.example.echo.domain.member.entity.Role
 import com.example.echo.domain.member.repository.MemberRepository
-import com.example.echo.domain.petition.dto.request.PetitionRequestDto
 import com.example.echo.domain.petition.entity.Category
 import com.example.echo.domain.petition.entity.Petition
 import com.example.echo.domain.petition.repository.PetitionRepository
@@ -11,10 +10,9 @@ import com.example.echo.global.exception.ErrorCode
 import com.example.echo.global.exception.PetitionCustomException
 import com.example.echo.log
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertNotEquals
+import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.assertThrows
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.data.domain.PageRequest
@@ -23,14 +21,12 @@ import org.springframework.test.context.TestPropertySource
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
 import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 @SpringBootTest
 @Transactional
 @TestPropertySource(locations = ["classpath:application-test.properties"])
-class PetitionServiceTests {
+class PetitionServiceUnitTests {
 
     @Autowired
     lateinit var petitionRepository: PetitionRepository
@@ -39,10 +35,12 @@ class PetitionServiceTests {
     lateinit var petitionService: PetitionService
 
     @Autowired
-    lateinit var memberRepository: MemberRepository
+    lateinit var memberRepository : MemberRepository
+
+    @Autowired
+    lateinit var summarizationService: SummarizationService
 
     private var admin: Member? = null
-    private var request: PetitionRequestDto? = null
 
     @BeforeEach
     fun setUp() {
@@ -69,98 +67,24 @@ class PetitionServiceTests {
                 likesCount = i
             })
         }
-
-        request = PetitionRequestDto(
-            memberId = admin!!.memberId!!,
-            title = "청원 요청 제목 테스트",
-            content = "청원 요청 내용 테스트",
-            startDate = LocalDateTime.now(),
-            endDate = LocalDateTime.now().plusDays(30),
-            category = Category.POLITICS,
-            originalUrl = "https://petitions.sample/0"
-        )
     }
 
     @Test
-    @DisplayName("관리자가 청원을 등록할 경우, 청원 데이터는 총 6개")
-    fun createPetitionTest() {
-        val savedPetition = petitionService.createPetition(request!!)
+    @DisplayName("청원 내용 요약 기능이 적절한 요약 결과를 반환하는지 검증")
+    fun getSummarizedTextTest() {
+        val content = """
+                국토교통부는 생활숙박시설에 대하여 주거규제를 추진함에 따라 2021년 5월 건축법 시행령을 개정하여 개정이후 분양되는 건축물 및 기존 생활숙박시설 건축물(소급적용) 전부에 대하여 숙박업 신고의무를 부과하고 주거사용시 건축물의 용도위반에 해당하여 이행강제금을 부과할 수 있도록 하는 정책을 발표하였습니다. 이에 따라 국토부는 2021년 10월 14일부터 2023년 10월 14일까지 생활숙박시설의 오피스텔 용도변경 유도정책을 추진한다고 하였으나 실제 용도변경을 완성한 곳은 1% 남짓에 불과하였습니다.
 
-        assertEquals("청원 요청 제목 테스트", savedPetition.title)
-        assertEquals(6, petitionRepository.findAll().size)
-    }
+                이러한 상황으로 인하여 생활숙박시설에 거주하고 있는 소유자, 주거 이용 목적으로 분양을 받은 수분양자들은 이행강제금을 부담할 수도 없고 대체 주거시설을 마련할 경제적 여유도 없기에 주거권을 박탈당할 위기에 놓였으며, 생계를 위협받으며 가정이 해체될 위험한 지경에 이르고 있습니다. 또한 건축 중인 생활숙박시설단지들 역시 수분양권자들은 주거가능 시설로 오인하게 하여 분양한 시행사를 상대로 분양권 취소 소송을 할 수 밖에 없고, 이로 인해 중도금 상환 연체, 잔급 미납 등으로 개인 신용불량 위험에 놓이며, 시행사와 시공사, 부동산 금융기관 역시 줄줄이 파산의 위협과 공포에 놓인 상황입니다. 오피스텔 용도변경을 추진하고 있는 생활숙박시설 단지의 관할 지방자치단체 역시 용도변경에 관련된 국토부의 종합적인 지침도 없이  적극 행정조치를 하지 못하고 민원을 해결할 수 없어 지방행정에 대한 불만누적으로 고통받고 있으며 현실에 맞지 않는 모순된 생활숙박시설 주거규제는 일파만파 사회적 문제를 야기하고 있습니다. 
 
-    @Test
-    @DisplayName("청원 단건 조회 - 요약이 없는 경우 요약 생성해서 반환")
-    fun getPetitionByIdNoSummaryTest() {
-        val petition = petitionRepository.save(Petition().apply {
-            member = admin
-            title = "청원 제목 테스트"
-            content = "청원 내용 테스트"
-            startDate = LocalDateTime.now()
-            endDate = LocalDateTime.now().plusDays(30L)
-            category = Category.POLITICS
-            originalUrl = "https://petitions.sample/i"
-            likesCount = 30
-        })
+                따라서 현실적이고 근본적인 문제 해결을 위해서는 현행 주택법 시행령에 규정에 되어 있는 준주택 제도를 활용하여 사실상 주거시설로 사용가능하고 사용되어지고 있는 생활숙박시설을 국민의 주거안정과 건축시설의 효율적인 이용을 위한 법상제도인 준주택으로 편입하여 현실에 맞는 제도정비를 통해 안정적인 주거생활로 가정을 유지할 수 있도록 노력해주실 것을 요청드리고자 청원에 이르게 되었습니다.
+            """.trimIndent()
 
-        assertEquals("청원 제목 테스트", petition.title)
-        assertNotEquals("요약은 비어있지 않아야 합니다.", "", petition.summary)
-        log.info("요약 결과: ${petition.summary}")
-    }
+        val summary = summarizationService.getSummarizedText(content)
 
-    @Test
-    @DisplayName("청원 단건 조회 - 요약이 있는 경우 요약 그대로 반환")
-    fun getPetitionByExistingSummaryTest() {
-        val summaryPetition = petitionRepository.save(Petition().apply {
-            member = admin
-            title = "요약 청원 제목 테스트"
-            content = "요약 청원 내용 테스트"
-            startDate = LocalDateTime.now()
-            endDate = LocalDateTime.now().plusDays(30L)
-            category = Category.POLITICS
-            originalUrl = "https://petitions.sample/10"
-            likesCount = 10
-            summary = "청원 요약 존재"
-        })
-
-        val foundPetition = petitionService.getPetitionById(summaryPetition.petitionId!!)
-
-        assertEquals("요약 청원 제목 테스트", foundPetition.title)
-        assertEquals("청원 요약 존재", foundPetition.summary)
-    }
-
-    @Test
-    @DisplayName("청원 단건 조회 - 해당 청원이 없는 경우 예외")
-    fun getPetitionByIdTestNotFound() {
-        val wrongPetitionId = 100L
-
-        val exception = assertThrows<PetitionCustomException> {
-            petitionService.getPetitionById(wrongPetitionId)
-        }
-
-        assertEquals(ErrorCode.PETITION_NOT_FOUND, exception.errorCode)
-    }
-
-    @Test
-    @DisplayName("청원 단건 조회 - 해당 청원 만료 기간이 지난 경우 예외")
-    fun getPetitionByIdTestExpired() {
-        val expiredPetition = petitionRepository.save(Petition().apply {
-            member = admin
-            title = "만료된 청원 제목 테스트"
-            content = "만료된 청원 내용 테스트"
-            startDate = LocalDateTime.now().minusDays(60L)
-            endDate = LocalDateTime.now().minusDays(30L)
-            category = Category.POLITICS
-            originalUrl = "https://petitions.sample/100"
-            likesCount = 100
-        })
-
-        val exception = assertThrows<PetitionCustomException> {
-            petitionService.getPetitionById(expiredPetition.petitionId!!)
-        }
-
-        assertEquals(ErrorCode.PETITION_EXPIRED, exception.errorCode)
+        assertNotEquals("summary는 비어있지 않아야 합니다.", "", summary)
+        assertTrue("summary는 4000자를 초과할 수 없습니다.") {summary.length <= 4000}
+        log.info("요약 결과: $summary")
     }
 
     @Test
@@ -248,36 +172,6 @@ class PetitionServiceTests {
     }
 
     @Test
-    @DisplayName("청원 좋아요 추가/제거 검증")
-    fun toggleLikeOnPetitionTest() {
-        val petition = petitionRepository.save(Petition().apply {
-            member = admin
-            title = "청원 제목 테스트 좋아요"
-            content = "청원 내용 테스트 좋아요"
-            startDate = LocalDateTime.now()
-            endDate = LocalDateTime.now().plusDays(30L)
-            category = Category.POLITICS
-            originalUrl = "https://petitions.sample/10"
-            likesCount = 4
-        })
-
-        assertEquals(4, petition.likesCount)
-        assertThat(petition.likedMemberIds).doesNotContain(admin!!.memberId!!)  // 좋아요 목록에 없는지 검증
-
-        val likeMessage = petitionService.toggleLikeOnPetition(petition.petitionId!!, admin!!.memberId!!)
-
-        assertEquals("좋아요가 추가되었습니다.", likeMessage)
-        assertEquals(5, petition.likesCount)
-        assertThat(petition.likedMemberIds).contains(admin!!.memberId!!) // 좋아요 추가 후 있는지 검증
-
-        val cancelMessage = petitionService.toggleLikeOnPetition(petition.petitionId!!, admin!!.memberId!!)
-
-        assertEquals("좋아요가 제거되었습니다.", cancelMessage)
-        assertEquals(4, petition.likesCount)
-        assertThat(petition.likedMemberIds).doesNotContain(admin!!.memberId!!)  // 좋아요 목록에 다시 없는지 검증
-    }
-
-    @Test
     @DisplayName("카테고리 선택 시 해당 카테고리 청원 5개 무작위 조회")
     fun getRandomCategoryPetitionsTest() {
         for (i in 0..9) {
@@ -300,38 +194,6 @@ class PetitionServiceTests {
     }
 
     @Test
-    @DisplayName("관리자 청원 수정")
-    fun updatePetitionTest() {
-        val petition = petitionRepository.save(Petition().apply {
-            member = admin
-            title = "청원 제목 테스트 수정"
-            content = "청원 내용 테스트 수정"
-            startDate = LocalDateTime.now()
-            endDate = LocalDateTime.now().plusDays(30L)
-            category = Category.POLITICS
-            originalUrl = "https://petitions.sample/10"
-            likesCount = 10
-        })
-
-        petitionService.updatePetition(petition.petitionId!!, request!!)
-
-        assertEquals("청원 요청 제목 테스트", petition.title)
-        assertNotEquals("청원 내용 테스트 0", petition.content, "새 요청 내용으로 변해야 합니다.")
-    }
-
-    @Test
-    @DisplayName("관리자 청원 수정 - 해당 청원이 없는 경우 예외 ")
-    fun updatePetitionTestNotFound() {
-        val wrongPetitionId = 1000L
-
-        val exception = assertThrows<PetitionCustomException> {
-            petitionService.updatePetition(wrongPetitionId, request!!)
-        }
-
-        assertEquals(ErrorCode.PETITION_NOT_FOUND, exception.errorCode)
-    }
-
-    @Test
     @DisplayName("관리자 청원 삭제")
     fun deletePetitionByIdTest() {
         val petition = petitionRepository.save(Petition().apply {
@@ -347,7 +209,7 @@ class PetitionServiceTests {
 
         petitionService.deletePetitionById(petition.petitionId!!)
 
-        val exception = assertThrows<PetitionCustomException> {
+        val exception = org.junit.jupiter.api.assertThrows<PetitionCustomException> {
             petitionService.getPetitionById(petition.petitionId!!)
         }
 
@@ -359,7 +221,7 @@ class PetitionServiceTests {
     fun deletePetitionByIdTestNotFound() {
         val wrongPetitionId = 100L
 
-        val exception = assertThrows<PetitionCustomException> {
+        val exception = org.junit.jupiter.api.assertThrows<PetitionCustomException> {
             petitionService.deletePetitionById(wrongPetitionId)
         }
